+++ packages/devtools/lib/src/timeline/timeline_screen.dart
@@ -19,7 +19,9 @@ import '../ui/elements.dart';
 import '../ui/icons.dart';
 import '../ui/material_icons.dart';
 import '../ui/primer.dart';
+import '../ui/service_extension_elements.dart';
 import '../ui/ui_utils.dart';
+import '../ui/vm_flag_elements.dart';
 import 'event_details.dart';
 import 'frames_bar_chart.dart';
 import 'timeline_controller.dart';
