+++ packages/devtools/lib/src/ui/ui_utils.dart
@@ -10,73 +10,13 @@ import 'package:meta/meta.dart';
 import '../framework/framework.dart';
 import '../globals.dart';
 import '../messages.dart';
-import '../service_extensions.dart';
-import '../service_registrations.dart';
-import '../utils.dart';
-import 'analytics.dart' as ga;
 import 'elements.dart';
 import 'environment.dart' as environment;
 import 'fake_flutter/dart_ui/dart_ui.dart';
 import 'html_icon_renderer.dart';
 import 'material_icons.dart';
-import 'primer.dart';
 
 const int defaultSplitterWidth = 10;
-
-CoreElement createExtensionCheckBox(
-    ToggleableServiceExtensionDescription extensionDescription) {
-  final extensionName = extensionDescription.extension;
-  final CoreElement input = checkbox();
-
-  serviceManager.serviceExtensionManager.hasServiceExtension(
-      extensionName, (available) => input.disabled = !available);
-
-  serviceManager.serviceExtensionManager.getServiceExtensionState(
-    extensionName,
-    (state) {
-      final InputElement e = input.element;
-      e.checked = state.value;
-    },
-  );
-
-  input.element.onChange.listen((_) {
-    final InputElement e = input.element;
-    serviceManager.serviceExtensionManager
-        .setServiceExtensionState(extensionName, e.checked, e.checked);
-  });
-  final inputLabel = label();
-  if (extensionDescription.icon != null) {
-    inputLabel.add(createIconElement(extensionDescription.icon));
-  }
-  inputLabel.add(span(text: extensionName));
-
-  final outerDiv = div(c: 'form-checkbox')
-    ..add(CoreElement('label')..add([input, inputLabel]));
-  input.setAttribute('title', extensionDescription.disabledTooltip);
-  return outerDiv;
-}
-
-List<CoreElement> getServiceExtensionElements() {
-  return [
-    div(c: 'btn-group collapsible-1200 nowrap margin-left')
-      ..add(<CoreElement>[
-        ServiceExtensionButton(performanceOverlay).button,
-        ServiceExtensionButton(slowAnimations).button,
-      ]),
-    div(c: 'btn-group collapsible-1200 nowrap margin-left')
-      ..add(<CoreElement>[
-        ServiceExtensionButton(debugPaint).button,
-        ServiceExtensionButton(debugPaintBaselines).button,
-      ]),
-    div(c: 'btn-group collapsible-1400 nowrap margin-left')
-      ..add(<CoreElement>[
-        ServiceExtensionButton(repaintRainbow).button,
-        ServiceExtensionButton(debugAllowBanner).button,
-      ]),
-    div(c: 'btn-group nowrap margin-left')
-      ..add(ServiceExtensionSelector(togglePlatformMode).selector)
-  ];
-}
 
 StatusItem createLinkStatusItem(
   CoreElement textElement, {
@@ -107,201 +47,6 @@ Future<void> maybeAddDebugMessage(Framework framework, String screenId) async {
       await serviceManager.connectedApp.isFlutterApp &&
       !await serviceManager.connectedApp.isProfileBuild) {
     framework.showMessage(message: debugWarning, screenId: screenId);
-  }
-}
-
-/// Button that calls a service extension.
-///
-/// Service extensions can be found in [service_extensions.dart].
-class ServiceExtensionButton {
-  ServiceExtensionButton(this.extensionDescription) {
-    button = PButton.icon(
-      extensionDescription.description,
-      extensionDescription.icon,
-      title: extensionDescription.disabledTooltip,
-    )..small();
-
-    final extensionName = extensionDescription.extension;
-
-    // Disable button for unavailable service extensions.
-    button.disabled = !serviceManager.serviceExtensionManager
-        .isServiceExtensionAvailable(extensionName);
-    serviceManager.serviceExtensionManager.hasServiceExtension(
-        extensionName, (available) => button.disabled = !available);
-
-    button.click(() => _click());
-
-    _updateState();
-  }
-
-  final ToggleableServiceExtensionDescription extensionDescription;
-  PButton button;
-
-  void _click() {
-    ga.select(extensionDescription.gaScreenName, extensionDescription.gaItem);
-
-    final bool wasSelected = button.element.classes.contains('selected');
-    serviceManager.serviceExtensionManager.setServiceExtensionState(
-      extensionDescription.extension,
-      !wasSelected,
-      wasSelected
-          ? extensionDescription.disabledValue
-          : extensionDescription.enabledValue,
-    );
-  }
-
-  void _updateState() {
-    // Select button whose state is already enabled.
-    serviceManager.serviceExtensionManager
-        .getServiceExtensionState(extensionDescription.extension, (state) {
-      final extensionEnabled = state.value == extensionDescription.enabledValue;
-      button.toggleClass('selected', extensionEnabled);
-      button.tooltip = extensionEnabled
-          ? extensionDescription.enabledTooltip
-          : extensionDescription.disabledTooltip;
-    });
-  }
-}
-
-/// Dropdown selector that calls a service extension.
-///
-/// Service extensions can be found in [service_extensions.dart].
-class ServiceExtensionSelector {
-  ServiceExtensionSelector(this.extensionDescription) {
-    selector = PSelect()
-      ..small()
-      ..clazz('on-page-selector')
-      ..change(_handleSelect)
-      ..tooltip = extensionDescription.tooltips.first ??
-          extensionDescription.description;
-
-    extensionDescription.displayValues.forEach(selector.option);
-
-    final extensionName = extensionDescription.extension;
-
-    // Disable selector for unavailable service extensions.
-    selector.disabled = !serviceManager.serviceExtensionManager
-        .isServiceExtensionAvailable(extensionName);
-    serviceManager.serviceExtensionManager.hasServiceExtension(
-        extensionName, (available) => selector.disabled = !available);
-
-    _updateState();
-  }
-
-  final ServiceExtensionDescription extensionDescription;
-
-  PSelect selector;
-
-  String _selectedValue;
-
-  void _handleSelect() {
-    if (selector.value == _selectedValue) return;
-
-    ga.select(extensionDescription.gaScreenName, extensionDescription.gaItem);
-
-    final extensionValue = extensionDescription
-        .values[extensionDescription.displayValues.indexOf(selector.value)];
-
-    serviceManager.serviceExtensionManager.setServiceExtensionState(
-      extensionDescription.extension,
-      true,
-      extensionValue,
-    );
-
-    _selectedValue = selector.value;
-  }
-
-  void _updateState() {
-    // Select option whose state is already enabled.
-    serviceManager.serviceExtensionManager
-        .getServiceExtensionState(extensionDescription.extension, (state) {
-      if (state.value != null) {
-        final selectedIndex = extensionDescription.values.indexOf(state.value);
-        selector.selectedIndex = selectedIndex;
-        _selectedValue = extensionDescription.displayValues[selectedIndex];
-      }
-    });
-  }
-}
-
-class SamplePeriodSelector {
-  SamplePeriodSelector() {
-    selector = PSelect()
-      ..small()
-      ..clazz('on-page-selector')
-      ..change(_handleSelect)
-      ..tooltip = 'The frequency at which the VM will collect CPU samples. For '
-          'a finer-grained profile, choose a smaller sample period. Please '
-          'read our documentation to understand the trade-offs associated with '
-          'this setting.'
-      ..option('Sample period: 1000 μs', value: '1000')
-      ..option('Sample period: 500 μs', value: '500')
-      ..option('Sample period: 250 μs', value: '250')
-      ..option('Sample period: 150 μs', value: '150')
-      ..option('Sample period: 50 μs', value: '50');
-
-    // Select 250 μs (the default profile period).
-    selector.selectedIndex = defaultSelectedIndex;
-  }
-
-  static const defaultSelectedIndex = 2;
-
-  PSelect selector;
-
-  String _selectedValue;
-
-  void setSamplePeriod() {
-    serviceManager.service.setFlag('profile_period', selector.value);
-  }
-
-  void _handleSelect() {
-    if (selector.value == _selectedValue) return;
-    setSamplePeriod();
-    _selectedValue = selector.value;
-  }
-}
-
-/// Button that calls a registered service from flutter_tools. Registered
-/// services can be found in [service_registrations.dart].
-class RegisteredServiceExtensionButton {
-  RegisteredServiceExtensionButton(
-    this.serviceDescription,
-    this.action,
-    this.errorAction,
-  ) {
-    button = PButton.icon(
-      serviceDescription.title,
-      serviceDescription.icon,
-      title: serviceDescription.title,
-    )
-      ..small()
-      ..hidden(true);
-
-    // Only show the button if the device supports the given service.
-    serviceManager.hasRegisteredService(
-      serviceDescription.service,
-      (registered) {
-        button.hidden(!registered);
-      },
-    );
-
-    button.click(() => _click());
-  }
-
-  final RegisteredServiceDescription serviceDescription;
-  final VoidAsyncFunction action;
-  final VoidFunctionWithArg errorAction;
-  PButton button;
-
-  void _click() async {
-    try {
-      button.disabled = true;
-      await action();
-    } catch (e) {
-      errorAction(e);
-    } finally {
-      button.disabled = false;
-    }
   }
 }
